split types to separate files
